be/requirements.txt
fastapi
sqlalchemy
pydantic
psycopg2-binary~=2.9.3
python-dotenv